refactor streamlit example
