+++ visualize/streamlit/loans-example/loans.py
@@ -7,57 +7,7 @@ import altair as alt
 import time
 from schedule import every, repeat, run_pending
 
-
-import configparser, os #re, json
-from snowflake.snowpark import Session
-from snowflake.snowpark.context import get_active_session
-
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.asymmetric import rsa
-from cryptography.hazmat.primitives.asymmetric import dsa
-from cryptography.hazmat.primitives import serialization
-
-def get_rsa_key(key_path):
-    with open(key_path, "rb") as key:
-        p_key= serialization.load_pem_private_key(
-            key.read(),
-            password=None,
-            backend=default_backend()
-        )
-
-    pkb = p_key.private_bytes(
-        encoding=serialization.Encoding.DER,
-        format=serialization.PrivateFormat.PKCS8,
-        encryption_algorithm=serialization.NoEncryption())
-
-    return pkb
-
-def getSession():
-    try:
-        return get_active_session()
-    except:
-        parser = configparser.ConfigParser()
-        filename = os.path.join(os.path.expanduser('~'), ".snowsql/config")
-        parser.read(filename)
-        section = "connections.dev"
-        auth = dict()
-
-        if 'password' in parser[section]:
-            auth =  {"password": parser.get(section, "password")}
-        elif 'private_key_path' in parser[section]:
-            key_path = parser.get(section, "private_key_path")
-            key = get_rsa_key(key_path)
-            auth =  {"private_key": key}
-        else:
-            st.write('Did not find password or key')
-
-        pars = {
-            "account": parser.get(section, "accountname"),
-            "user": parser.get(section, "username"),
-            **auth
-        }
-
-        return Session.builder.configs(pars).create()
+from database_connection import getSession
 
 def get_loan_data(table_type):
     session = getSession()
@@ -68,10 +18,14 @@ def get_loan_data(table_type):
 
     return data
 
+st.set_page_config(
+    page_title="Standard vs Dynamic Tables",
+    page_icon="🏂",
+    layout="wide")
 
-st.set_page_config(layout="wide")
-st.title("Standard vs Dynamic Tables :tada:")
+alt.themes.enable("dark")
 
+st.title("Loans by state :dollar:")
 
 with st.empty():
     @repeat(every(5).seconds)
